* Check if VNC has received screenshots. Return none if not (Bug #28674)
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.1/ucs-school@38348 1b283449-9f2f-0410-b571-8eb228e1a901

+++ ucs-school-umc-computerroom/umc/python/computerroom/__init__.py
@@ -365,10 +365,15 @@ class Instance( SchoolBaseModule ):
 		if not computer:
 			raise UMC_CommandError( 'Unknown computer' )
 
+		tmpfile = computer.screenshot
+		if tmpfile is None:
+			# vnc has not (yet) received any screenshots from the computer
+			# dont worry, try again later
+			self.finished( request.id, None )
+			return
 		response = Response( mime_type = MIMETYPE_JPEG )
 		response.id = request.id
 		response.command = 'COMMAND'
-		tmpfile = computer.screenshot
 		response.body = open( tmpfile.name ).read()
 		os.unlink( tmpfile.name )
 		self.finished( request.id, response )
